Przeorganizowanie projektu, usnięcie Node
[lab1] Punkt 3.1.12
